Update node resource documentation
Refined the Explorer urls under Mainnet, Testnet, and Devnet sections in the node resources documentation to specify EVM and Cosmos based explorers. Added more Cosmos explorers for Mainnet and one for Testnet. The changes provide users with a variety of options for each and respect the diversity of our blockchain ecosystems.