Update specific release version and statesync instructions
The "Running Node" tutorial documentation was updated. The link to download the binary was changed from the latest release to specifically version 11.0.7. Additionally, the Statesync instructions were shifted and elaborated to emphasize the need for replacing the binary with the latest release before the first run, and consequences if not followed. This revision provides more precise guidance, ensuring users follow best practices for node synchronization.