Refactor documentation structure
This commit represents a major restructuring of the Tenet project's documentation. Multiple files related to CLI and tutorials were deleted, moved or renamed for better organization. Relevance and consistency were the primary reasons for repositioning and renaming these articles in the documentation hierarchy.

This change aims to improve navigation and understanding for users reading the documentation. With this modification, the important concepts are easier to locate, essential tutorials are more straightforward, and the logical flow of topics is significantly improved.

Furthermore, a new file "deploy-contract.md" was added. It contains a hardhat configuration example to deploy contracts, thereby providing practical help for developers using the Tenet platform.

Deletions purely consisted of outdated or irrelevant information, ensuring that all the provided data is up-to-date and useful to users.