Add mempool documentation for validators
A new file named 'mempool.md' has been added to the 'validators' in 'node' section of the documentation. This file provides comprehensive information about the available mempool options in Tendermint. This includes explanations of FIFO and Prioritized Mempools, their implementations, configurations and useful resources for better understanding. This documentation is essential for validators to understand how transactions are processed, and the differences between the old FIFO and the new prioritized mempool implementations.