Add section on editing validator description in node doc
A new section has been added to the node tutorials documentation under 'running-node.md'. This section provides detailed instructions on how to edit the public description of a validator. The addition was necessary as it provides pertinent information to users on how to identify their validators and update certain parameters, thus enabling better stake decisions. This addition will improve user's understanding and navigation of the platform.