Update index page content
The content and links of the index page are updated to reflect current features and functionality. The cards now redirect to pages about DiPoS, incentivisation, native stablecoin, running a Tenet node, and becoming a validator. Additionally, the copies that summarise these topics have been revised for clarity and to ensure accuracy.