Add emojis to card labels
The commit enriches the visual representation of card labels on the website by adding emojis next to the text. The objective is to make labels more interactive and engaging for the users, enhancing their browsing experience on the page. List of modified labels includes "Learn about DiPoS", "Incentivisation", "Native Stablecoin", "Launch Your Tenet Node", "Become a Validator" and "Launch dApp on Tenet".